+++ code/nuvla/notifs/log.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import sys
 
 log_formatter = logging.Formatter(
@@ -11,9 +12,20 @@ logger.addHandler(stdout_handler)
 logger.setLevel(logging.INFO)
 
 
-def get_logger(who, level=logging.DEBUG):
+def loglevel_from_env(who):
+    who_env = f'{who.upper().replace("-", "_")}_LOGLEVEL'
+    if who_env not in os.environ:
+        return
+    try:
+        return logging._checkLevel(os.environ[who_env])
+    except ValueError as ex:
+        logging.info('Wrong log level provided for %s: %s', who_env, ex)
+        return
+
+
+def get_logger(who: str, level=logging.INFO):
     global stdout_handler
     log = logging.getLogger(who)
     log.addHandler(stdout_handler)
-    log.setLevel(level)
+    log.setLevel(loglevel_from_env(who) or level)
     return log